move search result into framework
