+++ LookupAnything/Components/SearchResultComponent.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using Pathoschild.LookupAnything.Framework;
 using StardewValley;
 using StardewValley.Menus;
